minor changes in the data keys dict
