+++ frappe/core/doctype/data_import/exporter.py
@@ -9,7 +9,7 @@ import frappe.permissions
 import re, csv, os
 from frappe.utils.csvutils import UnicodeWriter
 from frappe.utils import cstr, formatdate, format_datetime
-from  frappe.core.doctype.data_import.data_import import get_data_keys
+from frappe.core.doctype.data_import.importer import get_data_keys
 from six import string_types
 
 reflags = {
