+++ frappe/core/doctype/data_import/importer.py
@@ -16,8 +16,19 @@ from frappe.utils.dateutils import parse_date
 from frappe.utils.file_manager import save_url
 
 from frappe.utils import cint, cstr, flt, getdate, get_datetime, get_url
-from frappe.core.doctype.data_import.data_import import get_data_keys
 from six import text_type, string_types
+
+
+@frappe.whitelist()
+def get_data_keys():
+	return frappe._dict({
+		"data_separator": _('Start entering data below this line'),
+		"main_table": _("Table") + ":",
+		"parent_table": _("Parent Table") + ":",
+		"columns": _("Column Name") + ":",
+		"doctype": _("DocType") + ":"
+	})
+
 
 
 @frappe.whitelist()
